refactor(providers): split the OpenAI compatible provider by concern
The single file had grown past 390 lines. It is now a folder of three:
api.ts holds the pure request and response shaping with no obsidian
import and no plugin state, settings.ts holds the settings type, the
defaults, the upgrade-safe getter and the settings UI, provider.ts holds
the class that wires them to the plugin.

Behaviour is unchanged; the settings UI takes the connection check as a
callback so it does not depend on the class.
